Split dialog sections into smaller components
